requirements.txt
psycopg2-binary>=2.9.0
openai>=1.0.0